updated the algorithms code and reduced the cognative complexity, also fixed the speed select issue(the fast option was not working correctly) fixed it, and updated styling

+++ src/components/Nav.tsx
@@ -121,7 +121,7 @@ export function Nav({
           value={speed}
           options={SPEEDS}
           onChange={(e) => {
-            setSpeed(Number.parseInt(e.target.value) as SpeedType);
+            setSpeed(Number.parseFloat(e.target.value) as SpeedType);
           }}
         />
         <PlayButton
